refactor: carve agent prompt sections into modules (#4896)
