+++ server/services/promptSections/taskContext.js
@@ -1,0 +1,169 @@
+/**
+ * Shared task, attachment, and compaction prompt sections.
+ */
+
+import { join, basename } from 'path';
+import { PATHS } from '../../lib/fileUtils.js';
+import { TASK_PROMPT_KEY, TASK_CONTEXT_KEY } from '../../lib/cosTaskPrompt.js';
+import { PORTOS_APP_ID } from '../apps.js';
+
+/**
+ * Build a compaction instruction section for retries after context-limit failures.
+ * Provides explicit guidance to the agent on reducing output verbosity.
+ */
+export function buildCompactionSection(task) {
+  const compaction = task.metadata?.compaction;
+  if (!compaction?.needed) return '';
+
+  const hints = compaction.retryHints || [];
+  const reason = compaction.reason === 'output-limit' ? 'output length limit' : 'context window limit';
+  const prevOutputKB = compaction.outputSize ? Math.round(compaction.outputSize / 1024) : 'unknown';
+
+  return `
+## Context Compaction Required
+
+**WARNING**: A previous attempt at this task failed because the agent exceeded the ${reason}.
+Previous output size: ~${prevOutputKB} KB. You MUST keep your output compact to avoid the same failure.
+
+**Mandatory output constraints**:
+${hints.map(h => `- ${h}`).join('\n')}
+- Do NOT reproduce entire file contents in your output
+- Reference files by path and line number instead of quoting them
+- Limit exploratory reads — plan your approach first, then make targeted changes
+`;
+}
+
+/**
+ * Render a file-list task field (screenshots, attachments, …) either as a
+ * `### Header` + bulleted-path list (light path) or a single inline
+ * `**Header**: a, b` line (full path). Shared by every file-list field in
+ * `buildTaskBlock` so a wording tweak or a new field can't drift between them.
+ *
+ * @param {string} header - Section heading / inline label (e.g. "Screenshots").
+ * @param {Array} items - Task-metadata array; anything else renders as ''.
+ * @param {(item: any) => string} formatItem - Renders one item for the bulleted list.
+ * @param {(item: any) => string} formatInline - Renders one item for the inline join.
+ * @param {boolean} asList - Bulleted-list style vs. inline style.
+ * @returns {string}
+ */
+function renderFileListField(header, items, formatItem, formatInline, asList) {
+  if (!Array.isArray(items) || items.length === 0) return '';
+  return asList
+    ? `### ${header}\nUse your filesystem tools to inspect each path:\n` +
+      items.map(i => `- ${formatItem(i)}`).join('\n')
+    : `**${header}**: ${items.map(formatInline).join(', ')}`;
+}
+
+/**
+ * Build the shared task block — the description plus optional `**Target App**`,
+ * `**Screenshots**`, and `**Attachments**` fields. Used by BOTH the light and
+ * full prompt paths so a new task-metadata field gets surfaced in both without
+ * drift.
+ *
+ * Returns pre-rendered slots (`description`, `targetApp`, `screenshots`,
+ * `attachments`). Absent fields come back as empty strings so the full path's
+ * template literal can interpolate them in fixed positions and preserve
+ * byte-identical line spacing. The light path filters out the empty strings
+ * and joins what remains.
+ *
+ * @param {Object} task
+ * @param {Object} [opts]
+ * @param {boolean} [opts.screenshotsAsList=false] - When true, render screenshots
+ *   and attachments as a header followed by a bulleted list of paths (light
+ *   path style). When false, render as a single inline `**Field**: a, b`
+ *   line (full path style).
+ * @returns {{ description: string, targetApp: string, screenshots: string, attachments: string }}
+ */
+export function buildTaskBlock(task, { screenshotsAsList = false } = {}) {
+  const description = task.description;
+  // Only surface **Target App** for MANAGED apps — it scopes cross-repo work the
+  // agent's cwd wouldn't otherwise reveal. For the PortOS default app the agent
+  // already runs in the PortOS directory, so the line is redundant noise.
+  const app = task.metadata?.app;
+  const targetApp = app && app !== PORTOS_APP_ID ? `**Target App**: ${app}` : '';
+  const screenshots = renderFileListField(
+    'Screenshots', task.metadata?.screenshots,
+    (s) => `\`${resolveTaskFileRef(s)}\``, (s) => resolveTaskFileRef(s), screenshotsAsList
+  );
+  const label = (f) => (f?.originalName || f?.filename || f?.path || '').toString();
+  const path = (f) => resolveTaskFileRef((f?.path || '').toString());
+  const attachments = renderFileListField(
+    'Attachments', task.metadata?.attachments,
+    (f) => `\`${path(f)}\` (${label(f)})`, (f) => `${label(f)} (${path(f)})`, screenshotsAsList
+  );
+  return { description, targetApp, screenshots, attachments };
+}
+
+// Map for API-relative upload URLs → the on-disk root the agent should read.
+const TASK_REF_ROOTS = {
+  '/api/screenshots/': PATHS.screenshots,
+  '/api/attachments/': PATHS.cosAttachments,
+  '/api/uploads/': PATHS.uploads,
+};
+
+/**
+ * Resolve a stored screenshot/attachment reference to an absolute filesystem
+ * path the CoS agent can open with its filesystem tools.
+ *
+ * Uploads now return API-relative URLs (`/api/screenshots/<file>`) instead of
+ * absolute paths (issue #2518 — absolute paths leaked the install layout in the
+ * HTTP response), so tasks created after that change store the relative URL.
+ * Convert those back to an absolute path here, at prompt-build time, where the
+ * absolute path is local-only and never persisted or published. Legacy tasks
+ * that stored an absolute path (or any non-`/api/` value) are passed through
+ * unchanged so existing queues keep rendering the same path.
+ */
+function resolveTaskFileRef(ref) {
+  if (typeof ref !== 'string' || !ref) return ref;
+  for (const [prefix, root] of Object.entries(TASK_REF_ROOTS)) {
+    if (ref.startsWith(prefix)) {
+      return join(root, basename(decodeURIComponent(ref)));
+    }
+  }
+  return ref;
+}
+
+/**
+ * Undo the COS-TASKS.md round-trip split before rendering. The queue path in
+ * `cosTaskGenerator.js` persists a generated multi-line prompt by moving the
+ * full body into `metadata.prompt` (`metadata.context` before the #4153 split)
+ * and keeping only its first line as
+ * `description`, so the markdown stays one-line-per-task. Coming back out, the
+ * payload therefore *leads with a verbatim copy* of `description` — and every
+ * render path emits both (the task block, then the full body again under a
+ * `### Context` header). For a swarm task that surfaces as the reported
+ * double `# ⚡ SWARM MODE …` header; the same duplication hits any other
+ * generated/scheduled/system task that round-trips through the queue.
+ *
+ * When the split signature is present (the payload's first non-empty line
+ * equals `description`), fold it back into `description` so the prompt renders
+ * once, as one clean body with no spurious header. The split `prompt` key stays
+ * available to customized briefing templates; the legacy `context` payload key
+ * is dropped because it is already represented by `description`. A genuinely-
+ * separate user-supplied note (first line differs from `description`) is left
+ * untouched.
+ *
+ * Pure and idempotent: returns the same task when nothing matched, otherwise a
+ * shallow clone (never mutates the caller's task, so the stored task keeps its
+ * one-line `description` for the task-list UI).
+ */
+export function reconcileSplitContext(task) {
+  if (typeof task?.description !== 'string') return task;
+  const description = task.description.trim();
+  // `prompt` first: that is where the payload lives on a task written by
+  // current code (#4153). `context` stays in the loop so a task written before
+  // the split — or synced from a peer still on the old code — reconciles too.
+  // Keep the new prompt key available to customized templates. Drop only the
+  // legacy context key, whose payload is already represented by description.
+  for (const key of [TASK_PROMPT_KEY, TASK_CONTEXT_KEY]) {
+    const payload = task.metadata?.[key];
+    if (typeof payload !== 'string') continue;
+    // Mirror firstLine() in cosTaskStore.js: first non-empty, trimmed line.
+    const firstNonEmpty = payload.split('\n').map(l => l.trim()).find(Boolean) || '';
+    if (firstNonEmpty !== description) continue;
+    if (key === TASK_PROMPT_KEY) return { ...task, description: payload };
+    const { [key]: _dropped, ...restMeta } = task.metadata;
+    return { ...task, description: payload, metadata: restMeta };
+  }
+  return task;
+}
